+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   compiler, but drop the whole idea of webpack hot reloading. Not the most elegant
   but extremely simple to handle without much overhead. Also doesn't require
   any retooling (in fact allows deletion of dependencies).
+* Using `reload` refresh the browser on server or client code changes.